Fix quantization field access and SearchPlan initialization issues
This commit resolves the remaining compilation errors after quantized_vector field removal:

Issues Fixed:
1. Version type mismatches - Changed i64 to u32 in test files to match proto definition
2. Removed quantized_vector field access in viper pipeline tests (field no longer exists)
3. Added missing metadata_filters and query_vector fields to SearchPlan initializers
4. Updated SearchPlan initialization to use params.vector instead of non-existent params.query_vector

Files Modified:
- src/storage/engines/impls/sst/readers/tests/unified_sstable_reader_edge_tests.rs
- src/storage/engines/impls/viper/pipeline_tests.rs
- src/storage/engines/impls/viper/readers/tests/unified_parquet_reader_tests.rs
- src/storage/engines/impls/viper/tests/engine_tests.rs
- tests/integration/mvcc_logic_integration_test.rs

Compilation Status:
- Library builds successfully with 0 errors
- Reduced from 59 initial errors to 0 errors
- Ready for testing

+++ src/storage/engines/impls/sst/readers/tests/unified_sstable_reader_edge_tests.rs
@@ -510,7 +510,7 @@ mod edge_tests {
                 timestamp: version,
                 updated_at: Some(version),
                 expires_at: None,
-                version: Some(version as i64),
+                version: Some(version as u32),
                 source: None,
                 // level field removed from VectorRecord
             });
